refactor(trading-signals): rename HighLowClose.ts to Candle.types.ts
The file exports the whole candle type family (HighLow, HighLowClose,
OpenHighLowClose, OpenHighLowCloseVolume, HighLowCloseVolume), so naming
it after a single member undersold its contents. Type names and the
public API are unchanged.
