+++ packages/trading-signals/src/momentum/ER/ER.ts
@@ -1,5 +1,5 @@
 import {TradingSignal, TrendIndicatorSeries} from '../../types/Indicator.js';
-import type {HighLowClose} from '../../types/HighLowClose.js';
+import type {HighLowClose} from '../../types/Candle.types.js';
 import {getMaximum} from '../../util/getMaximum.js';
 import {getMinimum} from '../../util/getMinimum.js';
 import {pushUpdate} from '../../util/pushUpdate.js';
