+++ packages/trading-signals/src/trend/DX/DX.ts
@@ -2,7 +2,7 @@ import {ATR} from '../../volatility/ATR/ATR.js';
 import {IndicatorSeries} from '../../types/Indicator.js';
 import type {MovingAverage} from '../MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../MA/MovingAverageTypes.js';
-import type {HighLowClose} from '../../types/HighLowClose.js';
+import type {HighLowClose} from '../../types/Candle.types.js';
 import {WSMA} from '../WSMA/WSMA.js';
 
 /**
