+++ packages/trading-signals/src/trend/ZIGZAG/ZigZag.ts
@@ -1,5 +1,5 @@
 import {IndicatorSeries} from '../../types/Indicator.js';
-import type {HighLow} from '../../types/HighLowClose.js';
+import type {HighLow} from '../../types/Candle.types.js';
 
 export type ZigZagConfig = {
   /**
