+++ packages/trading-signals/src/types/Candle.types.ts
