+++ packages/trading-signals/src/types/index.ts
@@ -1,4 +1,4 @@
 export * from './BandsResult.js';
-export * from './HighLowClose.js';
+export * from './Candle.types.js';
 export * from './Indicator.js';
 export * from './Period.js';
