+++ packages/trading-signals/src/volatility/ATR/ATR.ts
@@ -2,7 +2,7 @@ import {IndicatorSeries} from '../../types/Indicator.js';
 import type {MovingAverage} from '../../trend/MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../../trend/MA/MovingAverageTypes.js';
 import {TR} from '../TR/TR.js';
-import type {HighLowClose} from '../../types/HighLowClose.js';
+import type {HighLowClose} from '../../types/Candle.types.js';
 import {WSMA} from '../../trend/WSMA/WSMA.js';
 
 /**
