+++ packages/trading-signals/src/volatility/TR/TR.ts
@@ -1,5 +1,5 @@
 import {IndicatorSeries} from '../../types/Indicator.js';
-import type {HighLowClose} from '../../types/HighLowClose.js';
+import type {HighLowClose} from '../../types/Candle.types.js';
 
 /**
  * True Range (TR)
