+++ packages/trading-signals/src/volume/AD/AD.ts
@@ -1,4 +1,4 @@
-import type {HighLowCloseVolume} from '../../types/HighLowClose.js';
+import type {HighLowCloseVolume} from '../../types/Candle.types.js';
 import {TradingSignal, TrendIndicatorSeries} from '../../types/Indicator.js';
 
 /**
