+++ packages/trading-signals/src/volume/VWMA/VWMA.ts
@@ -1,4 +1,4 @@
-import type {HighLowCloseVolume} from '../../types/HighLowClose.js';
+import type {HighLowCloseVolume} from '../../types/Candle.types.js';
 import {TradingSignal, TrendIndicatorSeries} from '../../types/Indicator.js';
 import type {MovingAverage} from '../../trend/MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../../trend/MA/MovingAverageTypes.js';
